pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clearwing"
version = "1.0.0"
description = "Autonomous vulnerability scanner and source-code hunter"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Eric Hartford", email = "eric@quixi.ai" },
]
maintainers = [
    { name = "Eric Hartford", email = "eric@quixi.ai" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Topic :: Security",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]

dependencies = [
    # Core
    "scapy>=2.5.0",
    "aiohttp>=3.8.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    # Network protocols
    "asyncssh>=2.13.0",
    "pysmb>=1.2.6",
    # Reporting
    "jinja2>=3.1.0",
    "reportlab>=4.0.0",
    "genai-pyo3>=0.1.9",
    "pydantic>=2.0.0",
    "docker>=7.0.0",
    # TUI
    "textual>=0.47.0",
    # Web UI
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "websockets>=11.0",
    # Token counting
    "tiktoken>=0.5.0",
    # Knowledge graph
    "networkx>=3.0",
    # Memory
    "numpy>=1.24.0",
    "chromadb>=1.0.0",
    # Source-hunt pipeline — callgraph extraction
    "tree-sitter>=0.21.0",
    "tree-sitter-c>=0.21.0",
    "tree-sitter-cpp>=0.21.0",
    "tree-sitter-python>=0.21.0",
    "tree-sitter-javascript>=0.21.0",
    "tree-sitter-go>=0.21.0",
    "tree-sitter-rust>=0.21.0",
]

[project.optional-dependencies]
metasploit = ["pymetasploit3>=1.0.0"]
browser = ["playwright>=1.40.0"]
vector = [
    "sentence-transformers>=2.2.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "ruff>=0.11.0",
    "build>=1.2.0",
    "twine>=6.0.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",
    "mkdocstrings[python]>=0.24.0",
]
all = [
    "clearwing[metasploit,browser,vector,dev,docs]",
]

[project.urls]
Homepage = "https://github.com/Lazarus-AI/clearwing"
Repository = "https://github.com/Lazarus-AI/clearwing"
Issues = "https://github.com/Lazarus-AI/clearwing/issues"
Documentation = "https://lazarus-ai.github.io/clearwing/"

[project.scripts]
clearwing = "clearwing:main"

[tool.setuptools.packages.find]
include = ["clearwing*"]

[tool.setuptools.package-data]
clearwing = [
    "py.typed",                              # PEP 561 — ship type info to consumers
    "core/config_data/*.yaml",
    "core/skills/vulnerabilities/*.md",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "strict"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "S", "C90"]
ignore = [
    "E501",   # line-too-long — deferred; many tool docstrings exceed 100
    "B008",   # function calls in default args
    "S101",   # assert — used in test-like code and invariant checks
    "S104",   # binding to all interfaces — intentional for proxy/callback servers
    "S108",   # insecure temp file usage — legitimate in sandbox/tool contexts
    "S110",   # try-except-pass — intentional in cleanup/teardown paths
    "S112",   # try-except-continue — intentional in error-tolerant loops
    "S310",   # audit URL open — security tool makes intentional URL requests
    "S311",   # random not for crypto — used for jitter/obfuscation variable names
    "S324",   # insecure hash — used for non-security purposes (checksums, dedup)
    "S603",   # subprocess without shell=True — core tool functionality
    "S607",   # partial executable path — intentional use of system tools
]

[tool.ruff.lint.mccabe]
max-complexity = 15

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B", "E402", "E501", "S"]
"clearwing/sourcehunt/runner.py" = ["C901"]
"clearwing/ui/commands/sourcehunt.py" = ["C901"]
"clearwing/ui/commands/interactive.py" = ["C901"]
"clearwing/ui/web/app.py" = ["C901"]
"clearwing/agent/runtime.py" = ["C901"]
"clearwing/agent/tools/hunt/analysis.py" = ["C901"]
"clearwing/agent/tools/hunt/discovery.py" = ["C901"]
"clearwing/sourcehunt/pool.py" = ["C901"]
"clearwing/sourcehunt/hunter.py" = ["C901"]
"clearwing/sourcehunt/preprocessor.py" = ["C901"]
"clearwing/sourcehunt/reporter.py" = ["C901"]
"clearwing/sourcehunt/campaign.py" = ["C901"]
"clearwing/llm/native.py" = ["C901"]
"clearwing/runners/parallel/executor.py" = ["C901"]
"clearwing/safety/audit/metrics.py" = ["C901"]
"clearwing/analysis/taint_tracker.py" = ["C901"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

# Scoped mypy gate, built up across phases. Every module here runs with
# `disallow_untyped_defs = true` and `warn_unused_ignores = true` — the
# typed-core policy.
#
#   - clearwing.findings.*          (Phase 3e)
#   - clearwing.capabilities        (Phase 4c)
#   - clearwing.sourcehunt.*        (Phase 5 — ~37 untyped-def errors
#                                    cleaned up via real annotations)
#   - clearwing.agent.tools.*       (Phase 5 — implicit Optional,
#                                    no-any-return, browser-state None
#                                    narrowing, and a handful of real
#                                    arg-type mismatches all fixed)
#   - clearwing.core.*              (Phase 5 — EventBus _handlers/_lock
#                                    class attributes, implicit Optional
#                                    on config/engine/logger/module_loader,
#                                    yaml stub ignore)
[[tool.mypy.overrides]]
module = [
    "clearwing.findings.*",
    "clearwing.capabilities",
    "clearwing.sourcehunt.*",
    "clearwing.agent.tools.*",
    "clearwing.core.*",
    "clearwing.agent.runtime",
    "clearwing.agent.protocols",
    "clearwing.providers.manager",
]
disallow_untyped_defs = true
warn_unused_ignores = true
